Move Templater tests to test namespace directly
These tests were formerly tests of the `Metalware::Templater` class, but
are no longer applicable to that now rendering has been moved to the
namespaces. They have been moved here (and slightly tweaked to still
work), since I think they may still have some value as they test
additional things to the existing tests, like the availability of repo
config values when templating. Keeping these in a separate `describe`
for now to avoid conflicts/interactions with the above, and since we may
just end up deleting/refactoring these away at some point.
