+++ spec/namespaces/alces_spec.rb
@@ -297,3 +297,101 @@ RSpec.describe Metalware::Namespaces::Alces do
     end
   end
 end
+
+RSpec.describe Metalware::Namespaces::Alces do
+  # These tests were formerly tests of the `Metalware::Templater` class, but
+  # are no longer applicable to that now rendering has been moved to the
+  # namespaces. They have been moved here (and slightly tweaked to still work),
+  # since I think they may still have some value as they test additional things
+  # to the above like the availability of repo config values when templating.
+  # Keeping these in a separate `describe` for now to avoid
+  # conflicts/interactions with the above, and since we may just end up
+  # deleting/refactoring these away at some point.
+  context 'old Templater tests' do
+    include AlcesUtils
+
+    let(:filesystem) do
+      FileSystem.setup do |fs|
+        fs.write template_path, template.strip_heredoc
+      end
+    end
+
+    # XXX Could adjust tests using this to only use template with parts they
+    # need, to make them simpler and less dependent on changes to this or each
+    # other.
+    let(:template) do
+      <<-EOF
+      This is a test template
+      some_passed_value: <%= domain.config.some_passed_value %>
+      some_repo_value: <%= domain.config.some_repo_value %>
+      erb_repo_value: <%= domain.config.erb_repo_value %>
+      very_recursive_erb_repo_value: <%= domain.config.very_recursive_erb_repo_value %>
+      nested.repo_value: <%= domain.config.nested ? domain.config.nested.repo_value : nil %>
+      EOF
+    end
+
+    let(:template_path) { '/template' }
+
+    def expect_renders(template_parameters, expected)
+      filesystem.test do |_fs|
+        # Strip trailing spaces from rendered output to make comparisons less
+        # brittle.
+        rendered = alces.render_file(
+          template_path, template_parameters
+        ).gsub(/\s+\n/, "\n")
+
+        expect(rendered).to eq(expected.strip_heredoc)
+      end
+    end
+
+    describe '#render_file' do
+      context 'without a repo' do
+        it 'renders template with no extra parameters' do
+          expected = <<-EOF
+          This is a test template
+          some_passed_value:
+          some_repo_value:
+          erb_repo_value:
+          very_recursive_erb_repo_value:
+          nested.repo_value:
+          EOF
+
+          expect_renders({}, expected)
+        end
+      end
+
+      context 'with repo' do
+        before do
+          filesystem.with_repo_fixtures('repo')
+        end
+
+        it 'renders template with repo parameters' do
+          expected = <<-EOF
+          This is a test template
+          some_passed_value:
+          some_repo_value: repo_value
+          erb_repo_value: repo_value
+          very_recursive_erb_repo_value: repo_value
+          nested.repo_value: nested_repo_value
+          EOF
+
+          expect_renders({}, expected)
+        end
+
+        context 'when template uses property of unset parameter' do
+          let(:template) do
+            'unset.parameter: <%= unset.parameter %>'
+          end
+
+          it 'raises' do
+            filesystem.test do
+              expect do
+                described_class.render_file(template_path, {})
+              end.to raise_error NameError
+            end
+          end
+        end
+      end
+    end
+  end
+end
